class-101/requirements.txt
asgiref==3.8.1
Django==5.2
django-model-utils==5.0.0
sqlparse==0.5.3
typing_extensions==4.13.1
